Add live activity attributes to widget extension
